+++ smart-contracts/cosmwasm-contract/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "cw-tpl-osmosis"
+name = "cw-client-example"
 version = "0.1.0"
-authors = ["Daniel Farina <dan@osmosis.team>"]
+authors = ["Harsh Mashru <h.mashru@supraoracles.com>"]
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
