+++ smart-contracts/cosmwasm-contract/src/integration_tests.rs
@@ -1,16 +1,15 @@
 #[cfg(test)]
 mod tests {
+    use crate::contract::{execute, instantiate, query};
     use crate::helpers::CwTemplateContract;
-    use crate::msg::InstantiateMsg;
+    use crate::msg::{ContractResponse, ExecuteMsg, InstantiateMsg, QueryMsg};
+    use cosmwasm_std::testing::{mock_dependencies_with_balance, mock_env, mock_info};
+    use cosmwasm_std::{coins, from_json};
     use cosmwasm_std::{Addr, Coin, Empty, Uint128};
     use cw_multi_test::{App, AppBuilder, Contract, ContractWrapper, Executor};
 
     pub fn contract_template() -> Box<dyn Contract<Empty>> {
-        let contract = ContractWrapper::new(
-            crate::contract::execute,
-            crate::contract::instantiate,
-            crate::contract::query,
-        );
+        let contract = ContractWrapper::new(execute, instantiate, query);
         Box::new(contract)
     }
 
@@ -34,7 +33,7 @@ mod tests {
         })
     }
 
-    fn proper_instantiate() -> (App, CwTemplateContract) {
+    fn test_proper_instantiate() -> (App, CwTemplateContract) {
         let mut app = mock_app();
         let cw_template_id = app.store_code(contract_template());
 
@@ -57,13 +56,13 @@ mod tests {
         (app, cw_template_contract)
     }
 
-    mod count {
+    mod update_contract {
         use super::*;
         use crate::msg::ExecuteMsg;
 
         #[test]
-        fn count() {
-            let (mut app, cw_template_contract) = proper_instantiate();
+        fn test_update_contract_integrated() {
+            let (mut app, cw_template_contract) = test_proper_instantiate();
 
             let msg = ExecuteMsg::UpdateSupraContract {
                 supra_pull_contract: "Test Update".to_string(),
@@ -71,5 +70,46 @@ mod tests {
             let cosmos_msg = cw_template_contract.call(msg).unwrap();
             app.execute(Addr::unchecked(USER), cosmos_msg).unwrap();
         }
+    }
+    #[test]
+    fn test_proper_initialization() {
+        let mut deps = mock_dependencies_with_balance(&coins(2, "token"));
+
+        let msg = InstantiateMsg {
+            supra_pull_contract: "Test".to_string(),
+        };
+        let info = mock_info("creator", &coins(1000, "earth"));
+
+        // we can just call .unwrap() to assert this was a success
+        let res = instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
+        assert_eq!(0, res.messages.len());
+
+        // it worked, let's query the state
+        let res = query(deps.as_ref(), mock_env(), QueryMsg::GetSupraPullContract {}).unwrap();
+        let value: ContractResponse = from_json(&res).unwrap();
+        assert_eq!("Test".to_string(), value.supra_pull_contract);
+    }
+
+    #[test]
+    fn test_update_supra_contract() {
+        let mut deps = mock_dependencies_with_balance(&coins(2, "token"));
+
+        let msg = InstantiateMsg {
+            supra_pull_contract: "Test".to_string(),
+        };
+        let info = mock_info("creator", &coins(2, "token"));
+        let _res = instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
+
+        // beneficiary can release it
+        let info = mock_info("anyone", &coins(2, "token"));
+        let msg = ExecuteMsg::UpdateSupraContract {
+            supra_pull_contract: "Test Update".to_string(),
+        };
+        let _res = execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+
+        // should increase counter by 1
+        let res = query(deps.as_ref(), mock_env(), QueryMsg::GetSupraPullContract {}).unwrap();
+        let value: ContractResponse = from_json(&res).unwrap();
+        assert_eq!("Test Update".to_string(), value.supra_pull_contract);
     }
 }
